feat(cli): SMI-5128 batch D — wrap audit/manage/telemetry telemetry (final CLI batch)
Wraps the remaining 12 CLI handlers with withTelemetry:
- audit-collisions: collisions
- audit: advisories + audit (parent uses commander's command.help(), the 3rd
  action arg, in place of the closure's audit.help())
- manage: list + update + remove
- telemetry: 6 subcommands (enable/disable/status/install-hook/uninstall-hook/
  reset-id) extracted to a new telemetry.action.ts sibling (439 LOC + 6 wraps
  would exceed the 500 gate); telemetry.ts keeps the factory + re-exports the
  run*/idTail test helpers. Mirrors the SMI-5127 sync.action.ts split.

CLI_DISPATCHER_MAP coverage 24 → 36 (all CLI command handlers now wrapped).
telemetry.test.ts (19 tests) green confirms the split preserved behavior + the
test re-export path. All files <500 (telemetry.ts 97, telemetry.action.ts 436).
